Reorganize component directories
