+++ App/components/index.ts
@@ -2,11 +2,11 @@
 
 import Button from './buttons/Button'
 import CheckBoxRow from './inputs/CheckBoxRow'
-import QRScanner from './inputs/QRScanner'
 import TextInput from './inputs/TextInput'
 import ContactListItem from './listItems/ContactListItem'
 import CredentialListItem from './listItems/CredentialListItem'
 import NotificationListItem from './listItems/NotificationListItem'
+import QRScanner from './misc/QRScanner'
 import HighlightTextBox from './texts/HighlightTextBox'
 import InfoTextBox from './texts/InfoTextBox'
 import Label from './texts/Label'
