+++ App/components/misc/QRScanner.tsx
